% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{estimate.expression}
\alias{estimate.expression}
\title{Train the model.}
\usage{
estimate.expression(model, cnt, library.size = NULL)
}
\arguments{
\item{model}{A trained model, which is a list returned by train().}

\item{cnt}{A matrix of read counts. Each row and column correspond to a gene and sample, respectively.}

\item{library.size}{A vector of library sizes (i.e. total read count) of samples. If Null is given, 
colSums(cnt) is automatically assigned. Default NULL.}
}
\value{
A list, where the estimated count matrix is named as "cnt" and the log-transformed read-per-million value is 
   named as "rpm."
}
\description{
Train the model.
}
\examples{
\dontrun{
ex <- nesage::estimate.expression(model, cnt)
}
}
