% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesage.R
\docType{package}
\name{nesage-package}
\alias{nesage}
\alias{nesage-package}
\title{nesage: Negative-binomial Sparse Additive Generative Model for shallow or single-cell RNA-Seq}
\description{
Provides functionality for inference of topics, which are latent structure of RNA-Seq data,
and estimation of gene expression.
}
\author{
\strong{Maintainer}: Koji Iwayama \email{fieldtranscriptome@gmail.com}

}
